Drag events now report the button that started them. Fixed bug preventing multiple systems from being defined in a module-definition. Added new saddle tiles. Added drag to terrain-raise and -lower, and added terrain-level. Removed unneeded Gradle dependencies.

+++ core/src/main/groovy/org/snowjak/city/service/GameService.groovy
@@ -118,7 +118,7 @@ class GameService {
 		for(def entity : entities) {
 			def component = (IsMapCell) entity.remove(IsMapCell)
 			if(component && map)
-				map.setEntity component.cellX, component.cellY, null
+				map.setEntity( (int) component.cellX, (int) component.cellY, null )
 			
 			state.engine.removeEntity entity
